remove unused files
Currently, These spec files are useless and have not been submitted
to OBS for building, so it's necessary to clean them up.
Without this change, `gbs build` results will contails build error
with dependency issue, and the build result will be much cleaner after
applying this patch.

Refer to the same issue in platform/upstream/gcc48:
https://review.tizen.org/gerrit/#/c/21357

Change-Id: I9f489740127f8633fbaa38b1367052f773be4fd9
Signed-off-by: Li Jinjing <jinjingx.li@intel.com>