Revert "rename info files"
This reverts commit d172df8002c87b7cd6a1cd91b33b106ca0c49d20.

This patch fixes the bootstrap failure.
If you perform the followinf actions:

	mkdir build
	cd build
	../configure
	make

the build will fail with the following error:

	cp libgomp.info libgomp.info
	make  all-recursive
	cp: `libgomp.info' and `libgomp.info' are the same file

Thus it seems that it would be better to revert the patch that is not
included to upstream.

Also the packagin section in spec-file was modified in order to fit
proper names of info files.

Change-Id: I1a053f7644902825e76a6ff4a2b4651830d373bb
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>