Revert "gcc: gcc dir version"
This patch was created long time ago. Nowadays version of GCC cannot be fully bootstrapped & regtested in the independent environment (i. e. outside GBS or OBS chroot). But we need to bootstrap & regtest so that to be sure that or new changes in compiler do not cause new regressions. This is a requirement of open source GCC community and we need to fulfill it if we want to submit any patches to upstream.

This reverts commit d2c4b00dc3a12f0563fd29933c6ff93682bd4910.

Change-Id: I61a0c30d62b1d6f0fe6dc227a5be37afcd508ec4
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>